Dev summoner api

Approved-by: Charlie Stokes <cstokes873@gmail.com>

+++ SummonerAPI/Summoner.Models/RiotApi/SummonerDto.cs
@@ -5,8 +5,8 @@
         public int ProfileIconId { get; set; }
         public string Name { get; set; }
         public string Puuid { get; set; }
-        public int SummonerLevel { get; set; }
-        public int RevisionDate { get; set; }
+        public long SummonerLevel { get; set; }
+        public long RevisionDate { get; set; }
         public string Id { get; set; }
         public string AccountId { get; set; }
     }
